requirements.txt
robotframework
robotframework-appiumlibrary
robotframework-faker
requests
robotframework-jsonlibrary
robotframework-requests
